+++ my-app/src/components/forms/formElements/checkboxInput.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
 import { InputProps } from 'types/types';
 
-export default class Checkbox extends React.Component<InputProps> {
-  render() {
-    const { label, inputRef } = this.props;
-    return (
-      <div className="input-checkbox">
-        <input type="checkbox" id={label} ref={inputRef} />
-        <label className="input-title" htmlFor={label}>
-          {label}
-        </label>
-      </div>
-    );
-  }
+export default function CheckboxInput(props: InputProps) {
+  const { label, inputRef } = props;
+  return (
+    <div className="input-checkbox">
+      <input type="checkbox" id={label} ref={inputRef} />
+      <label className="input-title" htmlFor={label}>
+        {label}
+      </label>
+    </div>
+  );
 }
