+++ my-app/src/components/forms/formElements/dateInput.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
 import { InputProps } from 'types/types';
 
-export default class DateInput extends React.Component<InputProps> {
-  render() {
-    const { label, inputRef } = this.props;
-    return (
-      <div className="input-date">
-        <label className="input-title" htmlFor={label}>
-          {label}
-        </label>
-        <input type="date" id={label} ref={inputRef} />
-      </div>
-    );
-  }
+export default function DateInput(props: InputProps) {
+  const { label, inputRef } = props;
+  return (
+    <div className="input-date">
+      <label className="input-title" htmlFor={label}>
+        {label}
+      </label>
+      <input type="date" id={label} ref={inputRef} />
+    </div>
+  );
 }
